+++ lib/geordi/commands/chromedriver_update.rb
@@ -1,9 +1,9 @@
 desc 'chromedriver-update', 'Update the chromedriver'
-
 long_desc <<-LONGDESC
 Example: `geordi chromedriver_update`
 
-This command will find and install the matching chromedriver for the currently installed Chrome.
+This command will find and install the matching chromedriver for the currently
+installed Chrome.
 LONGDESC
 
 def chromedriver_update