+++ lib/geordi/commands/setup.rb
@@ -6,17 +6,8 @@ Check out a repository and cd into its directory. Then let `setup` do the tiring
 work: run `bundle install`, create `database.yml`, create databases, migrate
 (all if applicable).
 
-If a local bin/setup file is found, Geordi skips these steps runs bin/setup
-for setup instead.
-
-After setting up, loads a remote database dump into the development db when
-called with the `--dump` option:
-
-    geordi setup -d staging
-
-After setting up, runs all tests when called with the `--test` option:
-
-    geordi setup -t
+If a local bin/setup file is found, Geordi skips its routine and runs bin/setup
+instead.
 LONGDESC
 
 option :dump, type: :string, aliases: '-d', banner: 'TARGET',