+++ lib/geordi/commands/update.rb
@@ -3,17 +3,6 @@ long_desc <<-LONGDESC
 Example: `geordi update`
 
 Performs: `git pull`, `bundle install` (if necessary) and migrates (if applicable).
-
-After updating, loads a dump into the development db when called with the
-`--dump` option:
-
-    geordi update -d staging
-
-After updating, runs all tests when called with the `--test` option:
-
-    geordi update -t
-
-See `geordi help update` for details.
 LONGDESC
 
 option :dump, type: :string, aliases: '-d', banner: 'TARGET',